+++ tests/cli/test_manifest.py
@@ -2,7 +2,7 @@ from spinta.testing.cli import SpintaCliRunner
 from spinta.manifests.tabular.helpers import striptable
 from spinta.testing.tabular import create_tabular_manifest
 from spinta.testing.tabular import load_tabular_manifest
-from spinta.testing.tabular import render_tabular_manifest
+from spinta.manifests.tabular.helpers import render_tabular_manifest
 
 
 def test_copy(rc, cli: SpintaCliRunner, tmpdir):