+++ tests/cli/test_show.py
@@ -1,0 +1,31 @@
+from spinta.testing.cli import SpintaCliRunner
+from spinta.manifests.tabular.helpers import striptable
+from spinta.testing.tabular import create_tabular_manifest
+from spinta.testing.workarounds import fix_s3_backend_issue
+
+
+def test_show(rc, cli: SpintaCliRunner, tmpdir):
+    manifest = striptable('''
+    id | d | r | b | m | property | type   | ref     | source      | prepare | level | access    | uri | title | description
+       | datasets/gov/example     |        |         |             |         |       | protected |     |       |
+       |   | data                 | memory | null    |             |         |       | protected |     |       |
+       |                          |        |         |             |         |       |           |     |       |
+       |   |   |   | Country      |        | code    | salis       |         |       | open      |     |       |
+       |   |   |   |   | code     | string |         | kodas       |         |       | public    |     |       |
+       |   |   |   |   | name     | string |         | pavadinimas |         |       | open      |     |       |
+       |                          |        |         |             |         |       |           |     |       |
+       |   |   |   | City         |        | name    | miestas     |         |       | open      |     |       |
+       |   |   |   |   | name     | string |         | pavadinimas |         |       | open      |     |       |
+       |   |   |   |   | country  | ref    | Country | salis       |         |       | open      |     |       |
+       |                          |        |         |             |         |       |           |     |       |
+       |   |   |   | Capital      |        | name    | miestas     |         |       | protected |     |       |
+       |   |   |   |   | name     | string |         | pavadinimas |         |       | protected |     |       |
+       |   |   |   |   | country  | ref    | Country | salis       |         |       | protected |     |       |
+    ''')
+
+    create_tabular_manifest(tmpdir / 'manifest.csv', manifest)
+
+    rc = fix_s3_backend_issue(rc)
+    result = cli.invoke(rc, ['show', tmpdir / 'manifest.csv'])
+
+    assert striptable(result.stdout) == manifest