+++ tests/test_inspect.py
@@ -7,8 +7,10 @@ from spinta.manifests.tabular.helpers import striptable
 from spinta.testing.cli import SpintaCliRunner
 from spinta.testing.config import configure
 from spinta.testing.datasets import Sqlite
+from spinta.testing.tabular import create_tabular_manifest
 from spinta.testing.tabular import load_tabular_manifest
-from spinta.testing.tabular import render_tabular_manifest
+from spinta.manifests.tabular.helpers import render_tabular_manifest
+from spinta.testing.workarounds import fix_s3_backend_issue
 
 
 def test_inspect(
@@ -37,37 +39,64 @@ def test_inspect(
       | rs                   | sql    |     | {sqlite.dsn} |
     ''')
 
-    cli.invoke(rc, ['inspect'])
+    cli.invoke(rc, ['inspect', '-o', tmpdir / 'result.csv'])
 
     # Check what was detected.
-    manifest = load_tabular_manifest(rc, tmpdir / 'manifest.csv')
+    manifest = load_tabular_manifest(rc, tmpdir / 'result.csv')
     manifest.datasets['dataset'].resources['rs'].external = 'sqlite'
     assert render_tabular_manifest(manifest) == striptable(f'''
-    id | d | r | b | m | property   | source     | prepare | type    | ref     | level | access    | uri | title | description
-       | dataset                    |            |         |         |         |       | protected |     |       |
-       |   | rs                     | sqlite     |         | sql     |         |       | protected |     |       |
-       |                            |            |         |         |         |       |           |     |       |
-       |   |   |   | Country        | COUNTRY    |         |         | id      |       | protected |     |       |
-       |   |   |   |   | id         | ID         |         | integer |         |       | protected |     |       |
-       |   |   |   |   | code       | CODE       |         | string  |         |       | protected |     |       |
-       |   |   |   |   | name       | NAME       |         | string  |         |       | protected |     |       |
-       |                            |            |         |         |         |       |           |     |       |
-       |   |   |   | City           | CITY       |         |         |         |       | protected |     |       |
-       |   |   |   |   | name       | NAME       |         | string  |         |       | protected |     |       |
-       |   |   |   |   | country_id | COUNTRY_ID |         | ref     | Country |       | protected |     |       |
+    id | d | r | b | m | property   | type    | ref     | source     | prepare | level | access    | uri | title | description
+       | dataset                    |         |         |            |         |       | protected |     |       |
+       |   | rs                     | sql     |         | sqlite     |         |       | protected |     |       |
+       |                            |         |         |            |         |       |           |     |       |
+       |   |   |   | Country        |         | id      | COUNTRY    |         |       | protected |     |       |
+       |   |   |   |   | id         | integer |         | ID         |         |       | protected |     |       |
+       |   |   |   |   | code       | string  |         | CODE       |         |       | protected |     |       |
+       |   |   |   |   | name       | string  |         | NAME       |         |       | protected |     |       |
+       |                            |         |         |            |         |       |           |     |       |
+       |   |   |   | City           |         |         | CITY       |         |       | protected |     |       |
+       |   |   |   |   | name       | string  |         | NAME       |         |       | protected |     |       |
+       |   |   |   |   | country_id | ref     | Country | COUNTRY_ID |         |       | protected |     |       |
     ''')
 
 
-def _fix_s3_backend_issue(rc: RawConfig) -> RawConfig:
-    # S3 backend is hardcoded globally and requires a fixture.
-    # This rc override deletes S3 backend from config.
-    return rc.fork({
-        'backends': {
-            'default': {
-                'type': 'memory',
-            }
-        },
+def test_inspect_from_manifest_table(
+    rc: RawConfig,
+    cli: SpintaCliRunner,
+    tmpdir: Path,
+    sqlite: Sqlite,
+):
+    # Prepare source data.
+    sqlite.init({
+        'COUNTRY': [
+            sa.Column('ID', sa.Integer, primary_key=True),
+            sa.Column('NAME', sa.Text),
+        ],
     })
+
+    create_tabular_manifest(tmpdir / 'manifest.csv', f'''
+    d | r | m | property     | type   | ref | source       | access
+    dataset                  |        |     |              |
+      | rs                   | sql    |     | {sqlite.dsn} |
+    ''')
+
+    cli.invoke(rc, [
+        'inspect', tmpdir / 'manifest.csv',
+        '-o', tmpdir / 'result.csv',
+    ])
+
+    # Check what was detected.
+    manifest = load_tabular_manifest(rc, tmpdir / 'result.csv')
+    manifest.datasets['dataset'].resources['rs'].external = 'sqlite'
+    assert render_tabular_manifest(manifest) == striptable(f'''
+    id | d | r | b | m | property | type    | ref | source  | prepare | level | access    | uri | title | description
+       | dataset                  |         |     |         |         |       | protected |     |       |
+       |   | rs                   | sql     |     | sqlite  |         |       | protected |     |       |
+       |                          |         |     |         |         |       |           |     |       |
+       |   |   |   | Country      |         | id  | COUNTRY |         |       | protected |     |       |
+       |   |   |   |   | id       | integer |     | ID      |         |       | protected |     |       |
+       |   |   |   |   | name     | string  |     | NAME    |         |       | protected |     |       |
+    ''')
 
 
 def test_inspect_format(
@@ -89,7 +118,7 @@ def test_inspect_format(
         ],
     })
 
-    rc = _fix_s3_backend_issue(rc)
+    rc = fix_s3_backend_issue(rc)
 
     cli.invoke(rc, [
         'inspect',
@@ -101,18 +130,18 @@ def test_inspect_format(
     manifest = load_tabular_manifest(rc, tmpdir / 'manifest.csv')
     manifest.datasets['dataset'].resources['sql'].external = 'sqlite'
     assert render_tabular_manifest(manifest) == striptable(f'''
-    id | d | r | b | m | property   | source     | prepare | type    | ref     | level | access    | uri | title | description
-       | dataset                    |            |         |         |         |       | protected |     |       |
-       |   | sql                    | sqlite     |         | sql     |         |       | protected |     |       |
-       |                            |            |         |         |         |       |           |     |       |
-       |   |   |   | Country        | COUNTRY    |         |         | id      |       | protected |     |       |
-       |   |   |   |   | id         | ID         |         | integer |         |       | protected |     |       |
-       |   |   |   |   | code       | CODE       |         | string  |         |       | protected |     |       |
-       |   |   |   |   | name       | NAME       |         | string  |         |       | protected |     |       |
-       |                            |            |         |         |         |       |           |     |       |
-       |   |   |   | City           | CITY       |         |         |         |       | protected |     |       |
-       |   |   |   |   | name       | NAME       |         | string  |         |       | protected |     |       |
-       |   |   |   |   | country_id | COUNTRY_ID |         | ref     | Country |       | protected |     |       |
+    id | d | r | b | m | property   | type    | ref     | source     | prepare | level | access    | uri | title | description
+       | dataset                    |         |         |            |         |       | protected |     |       |
+       |   | sql                    | sql     |         | sqlite     |         |       | protected |     |       |
+       |                            |         |         |            |         |       |           |     |       |
+       |   |   |   | Country        |         | id      | COUNTRY    |         |       | protected |     |       |
+       |   |   |   |   | id         | integer |         | ID         |         |       | protected |     |       |
+       |   |   |   |   | code       | string  |         | CODE       |         |       | protected |     |       |
+       |   |   |   |   | name       | string  |         | NAME       |         |       | protected |     |       |
+       |                            |         |         |            |         |       |           |     |       |
+       |   |   |   | City           |         |         | CITY       |         |       | protected |     |       |
+       |   |   |   |   | name       | string  |         | NAME       |         |       | protected |     |       |
+       |   |   |   |   | country_id | ref     | Country | COUNTRY_ID |         |       | protected |     |       |
     ''')
 
 
@@ -136,7 +165,7 @@ def test_inspect_cyclic_refs(
         ],
     })
 
-    rc = _fix_s3_backend_issue(rc)
+    rc = fix_s3_backend_issue(rc)
 
     cli.invoke(rc, [
         'inspect',
@@ -148,19 +177,19 @@ def test_inspect_cyclic_refs(
     manifest = load_tabular_manifest(rc, tmpdir / 'manifest.csv')
     manifest.datasets['dataset'].resources['sql'].external = 'sqlite'
     assert render_tabular_manifest(manifest) == striptable(f'''
-    id | d | r | b | m | property   | source     | prepare | type    | ref     | level | access    | uri | title | description
-       | dataset                    |            |         |         |         |       | protected |     |       |
-       |   | sql                    | sqlite     |         | sql     |         |       | protected |     |       |
-       |                            |            |         |         |         |       |           |     |       |
-       |   |   |   | City           | CITY       |         |         | id      |       | protected |     |       |
-       |   |   |   |   | id         | ID         |         | integer |         |       | protected |     |       |
-       |   |   |   |   | name       | NAME       |         | string  |         |       | protected |     |       |
-       |   |   |   |   | country_id | COUNTRY_ID |         | ref     | Country |       | protected |     |       |
-       |                            |            |         |         |         |       |           |     |       |
-       |   |   |   | Country        | COUNTRY    |         |         | id      |       | protected |     |       |
-       |   |   |   |   | id         | ID         |         | integer |         |       | protected |     |       |
-       |   |   |   |   | capital    | CAPITAL    |         | ref     | City    |       | protected |     |       |
-       |   |   |   |   | name       | NAME       |         | string  |         |       | protected |     |       |
+    id | d | r | b | m | property   | type    | ref     | source     | prepare | level | access    | uri | title | description
+       | dataset                    |         |         |            |         |       | protected |     |       |
+       |   | sql                    | sql     |         | sqlite     |         |       | protected |     |       |
+       |                            |         |         |            |         |       |           |     |       |
+       |   |   |   | City           |         | id      | CITY       |         |       | protected |     |       |
+       |   |   |   |   | id         | integer |         | ID         |         |       | protected |     |       |
+       |   |   |   |   | name       | string  |         | NAME       |         |       | protected |     |       |
+       |   |   |   |   | country_id | ref     | Country | COUNTRY_ID |         |       | protected |     |       |
+       |                            |         |         |            |         |       |           |     |       |
+       |   |   |   | Country        |         | id      | COUNTRY    |         |       | protected |     |       |
+       |   |   |   |   | id         | integer |         | ID         |         |       | protected |     |       |
+       |   |   |   |   | capital    | ref     | City    | CAPITAL    |         |       | protected |     |       |
+       |   |   |   |   | name       | string  |         | NAME       |         |       | protected |     |       |
     ''')
 
 
@@ -179,7 +208,7 @@ def test_inspect_self_refs(
         ],
     })
 
-    rc = _fix_s3_backend_issue(rc)
+    rc = fix_s3_backend_issue(rc)
 
     cli.invoke(rc, [
         'inspect',
@@ -191,14 +220,14 @@ def test_inspect_self_refs(
     manifest = load_tabular_manifest(rc, tmpdir / 'manifest.csv')
     manifest.datasets['dataset'].resources['sql'].external = 'sqlite'
     assert render_tabular_manifest(manifest) == striptable(f'''
-    id | d | r | b | m | property  | source    | prepare | type    | ref      | level | access    | uri | title | description
-       | dataset                   |           |         |         |          |       | protected |     |       |
-       |   | sql                   | sqlite    |         | sql     |          |       | protected |     |       |
-       |                           |           |         |         |          |       |           |     |       |
-       |   |   |   | Category      | CATEGORY  |         |         | id       |       | protected |     |       |
-       |   |   |   |   | id        | ID        |         | integer |          |       | protected |     |       |
-       |   |   |   |   | name      | NAME      |         | string  |          |       | protected |     |       |
-       |   |   |   |   | parent_id | PARENT_ID |         | ref     | Category |       | protected |     |       |
+    id | d | r | b | m | property  | type    | ref      | source    | prepare | level | access    | uri | title | description
+       | dataset                   |         |          |           |         |       | protected |     |       |
+       |   | sql                   | sql     |          | sqlite    |         |       | protected |     |       |
+       |                           |         |          |           |         |       |           |     |       |
+       |   |   |   | Category      |         | id       | CATEGORY  |         |       | protected |     |       |
+       |   |   |   |   | id        | integer |          | ID        |         |       | protected |     |       |
+       |   |   |   |   | name      | string  |          | NAME      |         |       | protected |     |       |
+       |   |   |   |   | parent_id | ref     | Category | PARENT_ID |         |       | protected |     |       |
     ''')
 
 
@@ -207,7 +236,7 @@ def test_inspect_oracle_sqldump_stdin(
     cli: SpintaCliRunner,
     tmpdir: Path,
 ):
-    rc = _fix_s3_backend_issue(rc)
+    rc = fix_s3_backend_issue(rc)
     cli.invoke(rc, [
         'inspect',
         '-r', 'sqldump', '-',
@@ -252,13 +281,13 @@ def test_inspect_oracle_sqldump_stdin(
     # Check what was detected.
     manifest = load_tabular_manifest(rc, tmpdir / 'manifest.csv')
     assert render_tabular_manifest(manifest) == striptable(f'''
-    id | d | r | b | m | property | source  | prepare | type    | ref | level | access    | uri | title | description
-       | dataset                  |         |         |         |     |       | protected |     |       |
-       |   | sqldump              | -       |         | sqldump |     |       | protected |     |       |
-       |                          |         |         |         |     |       |           |     |       |
-       |   |   |   | Country      | COUNTRY |         |         |     |       | protected |     |       |
-       |                          |         |         |         |     |       |           |     |       |
-       |   |   |   | City         | CITY    |         |         |     |       | protected |     |       |
+    id | d | r | b | m | property | type    | ref | source  | prepare | level | access    | uri | title | description
+       | dataset                  |         |     |         |         |       | protected |     |       |
+       |   | sqldump              | sqldump |     | -       |         |       | protected |     |       |
+       |                          |         |     |         |         |       |           |     |       |
+       |   |   |   | Country      |         |     | COUNTRY |         |       | protected |     |       |
+       |                          |         |     |         |         |       |           |     |       |
+       |   |   |   | City         |         |     | CITY    |         |       | protected |     |       |
     ''')
 
 
@@ -267,7 +296,7 @@ def test_inspect_oracle_sqldump_file_with_formula(
     cli: SpintaCliRunner,
     tmpdir: Path,
 ):
-    rc = _fix_s3_backend_issue(rc)
+    rc = fix_s3_backend_issue(rc)
     (tmpdir / 'dump.sql').write_text('''
     -- Šalys
     CREATE TABLE "GEO"."COUNTRY" (
@@ -286,9 +315,9 @@ def test_inspect_oracle_sqldump_file_with_formula(
     manifest = load_tabular_manifest(rc, tmpdir / 'manifest.csv')
     manifest.datasets['dataset'].resources['sqldump'].external = 'dump.sql'
     assert render_tabular_manifest(manifest) == striptable(f'''
-    id | d | r | b | m | property | source   | prepare                            | type    | ref | level | access    | uri | title | description
-       | dataset                  |          |                                    |         |     |       | protected |     |       |
-       |   | sqldump              | dump.sql | file(self, encoding: 'iso-8859-4') | sqldump |     |       | protected |     |       |
-       |                          |          |                                    |         |     |       |           |     |       |
-       |   |   |   | Country      | COUNTRY  |                                    |         |     |       | protected |     |       |
+    id | d | r | b | m | property | type    | ref | source   | prepare                            | level | access    | uri | title | description
+       | dataset                  |         |     |          |                                    |       | protected |     |       |
+       |   | sqldump              | sqldump |     | dump.sql | file(self, encoding: 'iso-8859-4') |       | protected |     |       |
+       |                          |         |     |          |                                    |       |           |     |       |
+       |   |   |   | Country      |         |     | COUNTRY  |                                    |       | protected |     |       |
     ''')